% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Server Logic for the Shiny App}
\usage{
server(input, output)
}
\arguments{
\item{input}{User input from the UI (e.g., year, price range, make, transmission, and body type).}

\item{output}{The output displayed in the app (e.g., plot and download handler).}
}
\value{
The reactive plot and download handler for the filtered data.
}
\description{
The server logic manages the interactive behavior of the Shiny app. It filters the dataset
based on user input and renders a histogram that shows the distribution of vehicle prices.
Additionally, users can download the filtered data as a CSV file.
}
