% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{User Interface (UI) for the Shiny App}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
The UI layout for the Shiny app.
}
\description{
The UI defines the layout and appearance of the Shiny app. It includes selectors for year, price range,
vehicle make, transmission type, and body type. A download button allows users to download the filtered dataset.
}
\section{Features}{

\itemize{
\item \strong{Select Year}: Choose the vehicle production year (2023, 2024, 2025).
\item \strong{Price Range Slider}: Adjust the price range to filter vehicles based on their price.
\item \strong{Select Make}: Choose one or more vehicle manufacturers.
\item \strong{Select Transmission Type}: Filter vehicles based on transmission type (e.g., Automatic, Manual).
\item \strong{Select Body Type}: Filter vehicles based on the body type (e.g., SUV, Sedan).
\item \strong{Download Button}: Download the filtered data as a CSV file.
}
}

\keyword{datasets}
