% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{VehiclePriceApp}
\alias{VehiclePriceApp}
\title{Run the Shiny Application}
\description{
This function runs the Shiny application for exploring vehicle prices from 2023 to 2025. It provides
interactive features to filter the data by year, price range, make, transmission, and body type. Users can
view the distribution of vehicle prices and download the filtered dataset as a CSV file.
}
\examples{
shinyApp(ui = ui, server = server)
}
