% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoPrice.R
\name{vehicles_clean}
\alias{vehicles_clean}
\title{vehicles_clean Dataset}
\format{
A data frame with n rows and the following m variables:
\describe{
\item{year}{Numeric. The production year of the vehicle (2023-2025).}
\item{price}{Numeric. The price of the vehicle in USD.}
\item{make}{Character. The manufacturer of the vehicle (e.g., Toyota, Ford).}
\item{model}{Character. The model name of the vehicle.}
\item{fuel}{Character. The type of fuel used (e.g., Gasoline, Diesel).}
\item{mileage}{Numeric. The mileage of the vehicle in miles per gallon.}
\item{transmission}{Character. The type of transmission (e.g., Automatic, Manual).}
\item{body}{Character. The type of vehicle body (e.g., SUV, Sedan, Pickup Truck).}
}
}
\source{
The dataset was sourced from Kaggle and contains vehicle data from 2023 to 2025.
}
\description{
A cleaned dataset of vehicle data containing information on vehicle prices, production year, and various technical attributes.
This dataset includes vehicles from 2023 to 2025 and is used to analyze vehicle pricing trends.
}
\examples{
data(vehicles_clean)
head(vehicles_clean)
}
