% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoPrice.R
\name{prepare_autoprice_dataset}
\alias{prepare_autoprice_dataset}
\title{Prepare AutoPrice Dataset}
\usage{
prepare_autoprice_dataset(file_path, overwrite = TRUE)
}
\arguments{
\item{file_path}{The file path to the raw dataset CSV file.}

\item{overwrite}{Logical. Whether to overwrite the existing \code{vehicles_clean} dataset.}
}
\value{
A cleaned dataset of vehicle data (\code{vehicles_clean}) is saved in the package’s data folder.
}
\description{
This function prepares the \code{AutoPrice} dataset by reading in raw vehicle data from a CSV file, removing missing values, and filtering out price outliers. The cleaned dataset is stored in the \code{vehicles_clean} object.
}
\details{
The cleaning process involves filtering out rows where the year or price is missing, and removing price outliers using the 1st and 99th quantiles.
}
\examples{
prepare_autoprice_dataset("data-raw/vehicles_dataset.csv", overwrite = TRUE)

}
